jetson/requirements-jetson.txt
# =============================================================================
# Clarity+ Jetson Nano Requirements
# =============================================================================
# PyTorch and torchvision MUST be installed separately from NVIDIA's Jetson
# wheels. Standard pip torch/torchvision are x86-only and will fail on ARM.
#
# Install PyTorch first (see JETSON.md for your JetPack version):
#   pip install torch torchvision  # from NVIDIA wheel URL for your JetPack
#
# Then: pip install -r requirements-jetson.txt
# =============================================================================

# Web framework
fastapi
uvicorn
python-multipart
requests

# Numeric & imaging
numpy
opencv-python-headless
Pillow

# Config
python-dotenv

# Face: InsightFace + ONNX Runtime (CPU or TensorRT on Jetson)
insightface
onnxruntime

# Pose & eyes: MediaPipe
mediapipe==0.10.9

# Skin: timm (requires torch/torchvision - install those first)
# Pin huggingface_hub<0.32 to avoid hf-xet -> puccinialin (fails on Jetson aarch64)
huggingface_hub<0.32
timm

# Voice service (mic, Vosk, TTS)
vosk>=0.3.42
sounddevice>=0.4.6
pyttsx3>=2.90
